+++ src/main.rs
@@ -1,8 +1,8 @@
-mod ast;
 mod context;
 mod errors;
 mod interner;
 mod lexer;
+mod parser;
 mod types;
 
 use context::{Context, Printable};
